metadata-service/requirements.txt
pika
pymongo
psycopg2-binary
mysql-connector-python